+++ src/Editor.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { EditorState } from 'prosemirror-state';
@@ -6,7 +6,7 @@ import { EditorView } from 'prosemirror-view';
 import { keydownHandler } from 'prosemirror-keymap';
 import { getPlugins } from './plugins';
 import { renderStatic, buildSchema } from './utils';
-import { useCollaborativeDoc } from './collaborative/useCollaborativeDoc';
+import { collabIsReady, sendCollabChanges } from './plugins/collaborative';
 
 require('./styles/base.scss');
 
@@ -43,72 +43,56 @@ const defaultProps = {
 	handleDoubleClick: undefined,
 };
 
-const createEditorState = (props, schema) =>
-	EditorState.create({
-		doc: schema.nodeFromJSON(props.initialContent),
-		schema: schema,
-		plugins: getPlugins(schema, {
-			customPlugins: props.customPlugins,
-			collaborativeOptions: props.collaborativeOptions,
-			onChange: props.onChange,
-			onError: props.onError,
-			initialContent: props.initialContent,
-			placeholder: props.placeholder,
-			isReadOnly: props.isReadOnly,
-		}),
-	});
-
-const createEditorView = (props, editorState, editorRef) =>
-	new EditorView(
-		{ mount: editorRef.current },
-		{
-			state: editorState,
-			editable: () => false,
-			handleKeyDown: keydownHandler({
-				// Block Ctrl-S from launching the browser Save window
-				'Mod-s': () => {
-					return true;
-				},
-			}),
-			handleClickOn: props.handleSingleClick,
-			handleDoubleClickOn: props.handleDoubleClick,
-		},
-	);
-
 const Editor = (props) => {
 	const editorRef = useRef();
 	const schema = useRef(null);
-	const [view, setView] = useState(null);
 
 	if (schema.current === null) {
 		schema.current = buildSchema(props.customNodes, props.customMarks, props.nodeOptions);
 	}
 
-	const collabDoc = useCollaborativeDoc({
-		firebaseRef: props.collaborativeOptions.firebaseRef,
-		editorView: view,
-		initialKey: props.collaborativeOptions.initialDocKey,
-		prosemirrorSchema: schema.current,
-	});
-
-	const isReadOnly = props.isReadOnly || collabDoc.preventEditing;
-
 	useEffect(() => {
-		if (view) {
-			view.setProps({
-				editable: () => !(collabDoc.preventEditing || props.isReadOnly),
-				dispatchTransaction: (transaction) => {
-					const newState = view.state.apply(transaction);
-					view.updateState(newState);
-					collabDoc.sendCollabChanges();
+		const state = EditorState.create({
+			doc: schema.current.nodeFromJSON(props.initialContent),
+			schema: schema.current,
+			plugins: getPlugins(schema.current, {
+				customPlugins: props.customPlugins,
+				collaborativeOptions: props.collaborativeOptions,
+				onChange: props.onChange,
+				onError: props.onError,
+				initialContent: props.initialContent,
+				placeholder: props.placeholder,
+				isReadOnly: props.isReadOnly,
+			}),
+		});
+
+		const view = new EditorView(
+			{ mount: editorRef.current },
+			{
+				state: state,
+				editable: (editorState) => {
+					return !props.isReadOnly && collabIsReady(editorState);
 				},
-			});
-		}
-	}, [view, isReadOnly, collabDoc.sendCollabChanges]);
-
-	useEffect(() => {
-		const editorState = createEditorState(props, schema.current);
-		setView(createEditorView(props, editorState, editorRef));
+				handleKeyDown: keydownHandler({
+					/* Block Ctrl-S from launching the browser Save window */
+					'Mod-s': () => {
+						return true;
+					},
+				}),
+				handleClickOn: props.handleSingleClick,
+				handleDoubleClickOn: props.handleDoubleClick,
+				dispatchTransaction: (transaction) => {
+					try {
+						const newState = view.state.apply(transaction);
+						view.updateState(newState);
+						sendCollabChanges(newState);
+					} catch (err) {
+						console.error('Error applying transaction:', err);
+						props.onError(err);
+					}
+				},
+			},
+		);
 	}, []);
 
 	/* Before createEditor is called from componentDidMount, we */