+++ src/collaborative/index.js
@@ -1,2 +1,0 @@
-export { default } from './collaborativeDoc';
-export { useCollaborativeDoc } from './useCollaborativeDoc';