+++ CHANGELOG.md
@@ -80,7 +80,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - fix: Do not show collaborative cursors if isReadOnly is set.
 
 ## [6.0.0] - 2019-07-15
-- update: collaborative plugin now fires an onUpdateLatestKey callback with the latest remote
+- update: collaborative plugin now fires an onUpdateHighestKey callback with the latest remote
 key from Firebase.
 - update: BREAKING - collaborationOptions updated to take ref, rather than firebase credentials
 - update: BREAKING - new firebase schema to support merges, branches, and other metadata.